Factor out common test setup code
